animateNoDamage failed to remove from current frame list
runRemainingAnimations is used to animate the available animation handle
which were not yet run, it actually animate those animation handle and
removes from the list by notifyAnimationsRan(). In case if these animation
handles are not removed, it leads to the Fatal Assert of runRemainingAnimations.

Some usecases causes the addition of animation handle when there is
no animators, in these cases this check prevents execution of
notifyAnimationsRan() leads to Fatal Assert crash of runRemainingAnimations.

Allowing the animation to happen irrespective of animators solves the problem.

Change-Id: Ifa72286cdc6ff785fcc999f3b951dd787e7003e9
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>